Move canvas code to canvas
